+++ Changelog.md
@@ -4,6 +4,7 @@
 
 - Implement HTTP/3.
 - Rename `quic.Cookie` to `quic.Token` and `quic.Config.AcceptCookie` to `quic.Config.AcceptToken`.
+- Distinguish between Retry tokens and tokens sent in NEW_TOKEN frames.
 
 ## v0.11.0 (2019-04-05)